backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.9
python-dotenv==1.0.1
cryptography==42.0.7
bcrypt==4.1.3
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
